Update app store URIs config structure
The app store URIs config has been changed from an object containing
fields for native and web URIs, to just a list of URIs to be tried in
order of precedence. If the initial app store that was used to install
the app can be resolved through one of the URIs, then it's given
priority, and attempted to launch directly. Otherwise, all the URIs are
attempted in order until one of them resolves to a registered Activity.
If none are found (this should only happen if they have been
uninstalled, as they're availability is checked as a prerequisite for
showing the update button), then a toast message is shown to that
effect.

There are also a few other small changes, such as removing the
CATEGORY_BROWSABLE restriction (as browsability does not seem to be a
core requirement), adding a FLAG_ACTIVITY_NEW_TASK flag to the intent,
updating the error message to be generic instead of referencing Google
Play, and a tweak to the phrasing in the javadocs.