Silently fetch and cache course structure of downloaded videos
For videos to appear correctly i.e. appear in the same order as they
appear in courseware, the course structure needs to be cached.
For example, if a user has downloaded multiple videos and then logs out,
we delete all the cached data. But upon logging back in, the downloaded
videos need to appear correctly which requires the videos' course's
structure.
This commit addresses the case mentioned above by first querying the
course ids from the downloaded videos that exist in our db and then
fetching and caching the course structure of those course upon login.