Update to Retrofit 2
This commit updates the Retrofit library from 1.9.0 to 2.1.0, which is
currently the latest release version. The API has been changed in a few
fundamental ways in version 2. The Retrofit converters have been moved
off into separate library projects, so we're now importing the Gson
converter as well. This also necessitated an update to Gson itself from
2.4 to 2.7 (the latest release version), as the converter uses new APIs
that were introduced in Gson version 2.6.

The main change in Retrofit 2 was the joining of the synchronous and
asynchronous APIs via the Call interface, which is what's returned from
the API methods along with the generic type of the class representing
the response body. Since all the methods already present an asynchronous
API, the various Task subclasses created specially for each API call
have been deleted in it's favor, except where multiple asynchronous
operations needed to be performed in sequence (this could be done with
RxJava once it's imported, without having to move everything in one
thread or creating a large nested chain of callbacks). The encapsulating
*API classes have also been removed as redundant where possible without
undue inconvenience, as they defeat the simple expressivity that's the
purpose of Retrofit. In cases where they passed specific parameters to
the actual API methods in the *Service classes, there parameters have
been added as part of the URL (where they can be resolved statically as
constants).

The error handling is also changed in Retrofit 2. RetrofitError has been
removed, and error codes returned by the server are delivered just like
standard responses to the same callback method, to be checked and
handled there manually. However, we don't have specific error responses
documented, and we were using a generic handler for these along with all
the other error scenarios. Nor do the specific callback implementations
need access to the whole response data other than the converted body
(though generic handlers may use this information to perform some
actions based on the headers etc.).

Therefore, generic base implementations of the new callback interface
have been created that define abstract methods according to the old
style, with all errors delivered to the failure callback method, and the
extra data (the Call and Response objects) stripped. The Callback class
in the http package does only this (for use in the new MVP framework
currently being used in the profiles module, which was using a basic
custom asynchronous executor and observer framework for it's HTTP
calls), while the ErrorHandlingCallback class is a replacement of the
functionality provided by the old Task framework that we were using
(providing callbacks for starting and ending, and error messages).

There was also an existing base implementation of the Retrofit callback
interface from version 1 named RetroApiCallback, which seems to have
copied the logic from an old version of the base Task implementation,
but since it needed a total redesign anyway, it was deleted (as I didn't
like it's name).

The HttpResponseStatusException has been retained to serve it's
original purpose, but the HttpConnectivityException class has been
removed in favor of IOException (which is now the error that Retrofit
delivers in the case of connectivity errors), along with the base
HttpException, as there doesn't seem to be any concrete and realized
benefit for having one base exception class for Retrofit errors.

It doesn't use weak references though, since that would have
necessitated the listeners to be strongly referenced in the view class,
or actually implemented by the main view (Activity or Fragment) class
itself. The latter is what's done in the Task implementations in
practice. However, that still doesn't help in the cases where it's
extended to update the view (thus holding a reference to it), usually in
the form of an anonymous inner subclass. Any generic solution to this
would be quite complex, and would have to be implemented at the base
view level as well to provide proper loose coupling, and this doesn't
seem to be a critical issue either (since the view is only retained
temporarily in memory, and we don't seem to be that tightly constrained
on it), so it's being deferred until such a time as the MVP framework is
implemented throughout the app (which already handles this situation
properly).

The Interceptor API has also been removed in Retrofit 2, and delegated
to OkHttp, which also added an interceptors API as of version 2.2.0.
Therefore all the Retrofit interceptors have now been converted to
OkHttp interceptors. The offline interceptor has also been registered in
the client that's returned by OkHttpUtil, based on a flag parameter that
is passed to the get method.